requirements.txt
# --- Google & YouTube API ---
google-api-python-client==2.165.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-auth==2.38.0
googleapis-common-protos==1.69.2
httplib2==0.22.0
uritemplate==4.1.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
pytube==15.0.0

# --- Core Dependencies ---
requests==2.32.3
python-dateutil==2.9.0.post0
setuptools==77.0.3

# --- Video/Image Processing ---
moviepy==1.0.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
opencv-python==4.11.0.86
pillow==11.1.0
proglog==0.1.10

# --- Data Handling ---
numpy==2.2.4
pandas==2.2.3
pytz==2025.1

# --- AI (Local & Cloud NLP) ---
openai==1.68.2
huggingface-hub==0.29.3
transformers==4.48.0
gpt4all==2.2.1
torch==2.6.0   # 🔸Keep only if GPT4All requires it

# --- Optional Helpers & Infra ---
PyYAML==6.0.2
Jinja2==3.1.6
tqdm==4.67.1
typing_extensions==4.12.2

# --- Pydantic (used in metadata_generator.py) ---
pydantic==2.10.6
pydantic_core==2.27.2
annotated-types==0.7.0

# --- Token & Auth Helpers ---
pyasn1==0.6.1
pyasn1_modules==0.4.1
proto-plus==1.26.1
protobuf==6.30.1
rsa==4.9
